+++ packages/devtools_extensions/example/foo/packages/foo/lib/src/foo_base.dart
@@ -2,10 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:convert';
-import 'dart:developer';
 import 'package:flutter/material.dart';
 
+/// This is an example widget that will be used from `app_that_uses_foo`.
 class FooBarBaz extends StatelessWidget {
   const FooBarBaz({super.key, this.onFoo, this.onBar, this.onBaz});
 
@@ -35,42 +34,4 @@ class FooBarBaz extends StatelessWidget {
       ),
     );
   }
-}
-
-class FooController {
-  FooController() {
-    if (!_initialized) {
-      initFoo();
-    }
-  }
-
-  static final _things = <String, String>{
-    '1': 'apple',
-    '2': 'banana',
-  };
-
-  static bool _initialized = false;
-
-  static void initFoo() {
-    registerExtension('ext.foo.getThing', (method, parameters) async {
-      final thingId = parameters['id'];
-      final thing = _things[thingId] ?? 'unknown thing';
-      final response = {
-        'id': thingId,
-        'value': thing,
-      };
-      return ServiceExtensionResponse.result(json.encode(response));
-    });
-    registerExtension('ext.foo.getAllThings', (method, parameters) async {
-      return ServiceExtensionResponse.result(json.encode(_things));
-    });
-
-    _initialized = true;
-  }
-
-  void addThing(int id, String thing) {
-    _things['$id'] = thing;
-  }
-
-  int get totalThings => _things.length;
 }